Split __init__.py so connector is now in seperate file.
